requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
botocore==1.34.69
can-decoder==0.1.9
etncanedge_files==0.0.2
canmatrix==1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
idna==3.7
isodate==0.6.1
J1939-PGN==0.4
jmespath==1.0.1
mdf-iter==2.1.1
multidict==6.0.5
numpy==1.24.4
pandas==2.0.3
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3fs==2024.3.1
six==1.16.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
wrapt==1.16.0
yarl==1.9.4

